refactor: move document stuff to a different module
