+++ connector-sdk/document/pom.xml
@@ -1,0 +1,19 @@
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>io.camunda.connector</groupId>
+    <artifactId>connector-sdk-parent</artifactId>
+    <relativePath>../pom.xml</relativePath>
+    <version>8.6.0-SNAPSHOT</version>
+  </parent>
+
+  <name>connector-document</name>
+  <description>Camunda Connector Document</description>
+  <artifactId>connector-document</artifactId>
+  <packaging>jar</packaging>
+
+
+
+</project>
