+++ connector-sdk/jackson-datatype-document/pom.xml
@@ -14,7 +14,7 @@
   <dependencies>
     <dependency>
       <groupId>io.camunda.connector</groupId>
-      <artifactId>connector-core</artifactId>
+      <artifactId>connector-document</artifactId>
     </dependency>
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
