+++ connector-sdk/jackson-datatype-document/src/main/java/io/camunda/connector/document/annotation/jackson/deserializer/DocumentDeserializer.java
@@ -18,10 +18,10 @@ package io.camunda.connector.document.annotation.jackson.deserializer;
 
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
-import io.camunda.connector.api.document.Document;
-import io.camunda.connector.api.document.DocumentFactory;
-import io.camunda.connector.api.document.operation.DocumentOperationExecutor;
 import io.camunda.connector.document.annotation.jackson.DocumentReferenceModel;
+import io.camunda.document.Document;
+import io.camunda.document.DocumentFactory;
+import io.camunda.document.operation.DocumentOperationExecutor;
 import java.io.IOException;
 
 public class DocumentDeserializer extends DocumentDeserializerBase<Document> {
