+++ connector-sdk/jackson-datatype-document/src/main/java/io/camunda/connector/document/annotation/jackson/deserializer/ObjectDocumentDeserializer.java
@@ -19,9 +19,9 @@ package io.camunda.connector.document.annotation.jackson.deserializer;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.deser.std.UntypedObjectDeserializer;
-import io.camunda.connector.api.document.DocumentFactory;
-import io.camunda.connector.api.document.operation.DocumentOperationExecutor;
 import io.camunda.connector.document.annotation.jackson.DocumentReferenceModel;
+import io.camunda.document.DocumentFactory;
+import io.camunda.document.operation.DocumentOperationExecutor;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
