+++ connector-sdk/pom.xml
@@ -20,6 +20,7 @@
     <module>core</module>
     <module>validation</module>
     <module>test</module>
+    <module>document</module>
     <module>feel-wrapper</module>
     <module>jackson-datatype-feel</module>
     <module>jackson-datatype-document</module>
